Pin change for servo
servo wasnt working on the previous pin due to adding the Lora shield.
Note: The componenten together draw a lot of energy together to the point that when the Servo actuates, the LCD struggles to keep its text well lit. We might need to look into adding an additional powersource or split things up among multiple arduino boards.